Do fluidify when transforming symbols instead of on access
